+++ lib/avro/builder/enum.rb
@@ -1,0 +1,8 @@
+module Avro
+  module Builder
+
+    class Enum < Avro::Builder::Types::EnumType
+      include Avro::Builder::Types::TopLevel
+    end
+  end
+end