+++ lib/avro/builder/fixed.rb
@@ -1,0 +1,8 @@
+module Avro
+  module Builder
+
+    class Fixed < Avro::Builder::Types::FixedType
+      include Avro::Builder::Types::TopLevel
+    end
+  end
+end