+++ lib/avro/builder/record.rb
@@ -1,0 +1,8 @@
+module Avro
+  module Builder
+
+    class Record < Avro::Builder::Types::RecordType
+      include Avro::Builder::Types::TopLevel
+    end
+  end
+end